refactor: location of the constants
moved shared/services/constants to shared/constants and deleted the services constants folder
