+++ src/modules/notes/Note/Note.tsx
@@ -3,7 +3,7 @@ import NoteView from "./NoteView";
 import { fetchGetNotesService } from "modules/notes/Note/services";
 import { useQuery } from "@tanstack/react-query";
 import { useParams } from "react-router-dom";
-import { NOTE_QUERY_KEY } from "shared/services/constants/notes";
+import { NOTE_QUERY_KEY } from "shared/constants";
 
 const Note: React.FC = () => {
 	const { board_id } = useParams();
