+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/DialogDeleteAllTexts.tsx
@@ -5,7 +5,7 @@ import { useContextText } from "modules/texts/Text/Context";
 import { useDialogText } from "modules/texts/Text/shared/hooks/useDialogText";
 import DialogDeleteAllTextView from "modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/DialogDeleteAllTextsView";
 import { fetchDeleteAllTextsService } from "modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/services";
-import { TEXT_QUERY_KEY } from "shared/services/constants/texts";
+import { TEXT_QUERY_KEY } from "shared/constants";
 
 const DialogDeleteAllTexts: React.FC = () => {
 	const { board_id } = useParams();
