+++ src/shared/components/Drawer/components/Aside/components/DialogNewBoard/DialogNewBoard.tsx
@@ -4,7 +4,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import React, { memo, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { ASIDE_QUERY_KEY } from "shared/components/Drawer/components/Aside/constants";
-import { BOARD_QUERY_KEY } from "shared/services/constants/dashboard";
+import { BOARD_QUERY_KEY } from "shared/constants";
 import DialogNewBoardView from "shared/components/Drawer/components/Aside/components/DialogNewBoard/DialogNewBoardView";
 import { fetchPostBoardService } from "shared/components/Drawer/components/Aside/components/DialogNewBoard/services";
 import { DialogNewBoardSchema } from "shared/components/Drawer/components/Aside/components/DialogNewBoard/schema";
