+++ src/shared/constants/index.ts
@@ -1,2 +1,4 @@
-export * as default from "./app.constants";
 export * from "./app.constants";
+export * from "./dashboard.constants";
+export * from "./notes.constants";
+export * from "./texts.constants";
