+++ src/shared/constants/notes.constants.ts
