+++ src/shared/constants/texts.constants.ts
