--- src/shared/services/constants/dashboard/index.ts
@@ -1,1 +1,0 @@
-export { BOARD_QUERY_KEY } from "./board.constants";
