--- src/shared/services/constants/notes/index.ts
@@ -1,1 +1,0 @@
-export { NOTE_QUERY_KEY } from "./notes.constants";
