--- src/shared/services/constants/texts/index.ts
@@ -1,1 +1,0 @@
-export { TEXT_QUERY_KEY } from "./text.constants";
